Feature/v2.0 jsons (#57)
* Bump lib version

* Update chains main links

* Adopt fixed ss58 codec

* Bump lib version to fix option decoding

+++ core-api/src/main/java/io/novafoundation/nova/core/model/RuntimeConfiguration.kt
@@ -5,5 +5,5 @@ package io.novafoundation.nova.core.model
 class RuntimeConfiguration(
     val genesisHash: String,
     val erasPerDay: Int,
-    val addressByte: Byte,
+    val addressByte: Short,
 )
